adding makefile for checker

+++ src/print_stack.c
@@ -19,7 +19,7 @@ int 	check_order(t_ivec *a, t_ivec *b)
 	if (b->length)
 		return (0);
 	i = -1;
-	while (++i < a->length - 1)
+	while ((size_t)++i < a->length - 1)
 		if (a->data[i] > a->data[i + 1])
 			return (0);
 	return (1);
@@ -31,12 +31,12 @@ void	print_stack(t_ivec *a, t_ivec *b)
 
 	i = -1;
 	ft_printf("\nstack a\n");
-	while (++i < a->length)
+	while ((size_t)++i < a->length)
 		ft_printf("|%d|", a->data[i]);
 	ft_printf("\n");
 	i = -1;
 	ft_printf("\nstack b\n");
-	while(++i < b->length)
+	while((size_t)++i < b->length)
 		ft_printf("|%d|", b->data[i]);
 	ft_printf("\n");
 }
